pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "time_config_hub"
version = "1.0.0"
authors = [
    { name = "Ahmad Tarmizi, Noor Azura", email = "noor.azura.ahmad.tarmizi@intel.com" },
    { name = "Koay, Eric Eng Keong", email = "eng.keong.koay@intel.com" },
    { name = "Liow, Shi Jie Donavan", email = "shi.jie.donavan.liow@intel.com" },
]
description = "Time Config Hub for Intel TSN-capable hardware"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pyyaml>=6.0",
    "lxml>=6.0",
    "click>=8.2",
    "netifaces>=0.11",
    "watchdog>=6.0",
    "colorama>=0.4.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4",
    "black==25.12.0",
    "ruff==0.14.8",
    "pylint==3.3",
]

[project.scripts]
tch = "time_config_hub.cli:main"

[project.urls]
Homepage = "https://github.com/intel/time-confighub"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
time_config_hub = ["templates/*.service"]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]


[tool.ruff]
target-version = "py39"
line-length = 88

# Keep exclusions aligned with Black.
extend-exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
# Ruff defaults are intentionally conservative; this makes the baseline explicit.
select = ["E", "F"]
ignore = ["E501"]


[tool.pylint.messages_control]
disable = ["W1203", "W1201"]

